M    utils.py + warn if the lockfile misteriously dissapears.

+++ psync/utils.py
@@ -183,7 +183,7 @@ def grab(filename, url, limit=0, cont=True, progress=False):
         logger.debug ("cec= 0x%x" % curlExitCode)
 
     if curlExitCode==2:
-	logger.debug ("curl C-c'ed!")
+        logger.debug ("curl C-c'ed!")
         raise KeyboardInterrupt
     elif curlExitCode==0x1700:
         e= IOError ()
@@ -196,7 +196,7 @@ def grab(filename, url, limit=0, cont=True, progress=False):
 def lockFile (path):
     ans= True
     try:
-	makedirs (dirname (path))
+        makedirs (dirname (path))
         os.open (path, os.O_CREAT|os.O_EXCL, 0400)
     except OSError, e:
         if e.errno==17:
@@ -208,6 +208,9 @@ def lockFile (path):
     
 def unlockFile (path):
     logger.debug ("unlocking %s" % path)
-    unlink (path)
+    if not stat (path):
+        logger.warn ("lock %s dissapeared!" % path)
+    else:
+        unlink (path)
 
 #end
